refactor(timestamp): change last_mod_date and last_modified_date from int64 to int
- Type definitions and Patch types:
  + SampleRef.last_modified_date (clip.ml)
  + PatchRef.last_mod_date (device.ml)
- XML parsing functions (get_int_attr instead of get_int64_attr)
- Diff functions (Equality.IntEq instead of Equality.Int64Eq)
- format_unix_timestamp to accept int instead of int64
- All related test code

+++ test/test_diff_audio_clip.ml
@@ -96,8 +96,8 @@ let test_diff_logic () =
 
       (match sample_ref_patch.last_modified_date with
       | `Modified m ->
-          check int64 "old last modified" 1742403846L m.oldval;
-          check int64 "new last modified" 1742403845L m.newval
+          check int "old last modified" 1742403846 m.oldval;
+          check int "new last modified" 1742403845 m.newval
       | _ -> fail "Expected last modified date to be modified")
   | `Unchanged -> fail "Expected sample reference to be modified")
